+++ sources/Interop/D3D12MemoryAllocator/src/D3D12MemAlloc/D3D12MA_BlockMetadata_TLSF.Manual.cs
@@ -1,6 +1,6 @@
 // Copyright © Tanner Gooding and Contributors. Licensed under the MIT License (MIT). See License.md in the repository root for more information.
 
-// Ported from D3D12MemAlloc.cpp in D3D12MemoryAllocator tag v3.0.1
+// Ported from D3D12MemAlloc.cpp in D3D12MemoryAllocator tag v3.1.0
 // Original source is Copyright © Advanced Micro Devices, Inc. All rights reserved. Licensed under the MIT License (MIT).
 
 using System.Runtime.CompilerServices;
@@ -654,7 +654,7 @@ internal unsafe partial struct D3D12MA_BlockMetadata_TLSF
     {
         if (pThis->m_AllocCount == 0)
         {
-            return (ulong)(0);
+            return 0;
         }
 
         for (Block* block = pThis->m_NullBlock->prevPhysical; block != null; block = block->prevPhysical)
@@ -666,7 +666,7 @@ internal unsafe partial struct D3D12MA_BlockMetadata_TLSF
         }
 
         D3D12MA_FAIL("If m_AllocCount > 0 then should find any allocation!");
-        return (ulong)(0);
+        return 0;
     }
 
     [VtblIndex(14)]
@@ -684,7 +684,7 @@ internal unsafe partial struct D3D12MA_BlockMetadata_TLSF
                 return (ulong)(block);
             }
         }
-        return (ulong)(0);
+        return 0;
     }
 
     [VtblIndex(15)]
